Add files via upload
class that implements MyMap using open addressing with linear probing, as described on p. 989.  
See the lecture on linear probing. 
For simplicity, use f(key) = key % size as the hash function, where size is the hash-table size. Initially, 
the hash-table size is 4. The table size is doubled whenever the load factor exceeds the threshold 
(0.5).